+++ react-static-web-apps-auth/src/types.ts
@@ -4,4 +4,5 @@ export type AuthProviders =
   | "github"
   | "google"
   | "twitter"
+  | "apple"
   | string;